帮派技能初始化数据

+++ server-api/src/main/java/ren/liushuang/mytool/serverapi/entity/RenwuxiulianEntity.java
@@ -11,8 +11,8 @@ import lombok.NoArgsConstructor;
 @Builder
 public class RenwuxiulianEntity {
     private int level;
-    private int xiulianExp;
+    private long xiulianExp;
     private XiulianType type;
-    private int coin;
-    private int coinTotal;
+    private long coin;
+    private long coinTotal;
 }
